+++ packages/core/src/components.d.ts
@@ -28,14 +28,13 @@ export namespace Components {
     }
     interface DsoAutosuggest {
         /**
-          * A method that will be called debounced with the input value as its first parameter. This method will also be called when the input is reduced to an empty string.
-          * @returns A promise with an array of `Suggestion`s. You should limit this array to ten items.
-         */
-        "fetchSuggestions"?: (value: string) => Promise<Array<Suggestion>>;
-        /**
           * Whether the previous suggestions will be presented when the input gets focus again.
          */
         "suggestOnFocus": boolean;
+        /**
+          * The suggestions for the value of the slotted input element
+         */
+        "suggestions": Suggestion[];
     }
     interface DsoBadge {
         "status"?: 'primary' | 'success' | 'info' | 'warning' | 'danger';
@@ -359,18 +358,21 @@ declare namespace LocalJSX {
     }
     interface DsoAutosuggest {
         /**
-          * A method that will be called debounced with the input value as its first parameter. This method will also be called when the input is reduced to an empty string.
-          * @returns A promise with an array of `Suggestion`s. You should limit this array to ten items.
+          * This is emitted debounced for every change for the slotted input type=text element.
          */
-        "fetchSuggestions"?: (value: string) => Promise<Array<Suggestion>>;
+        "onDsoChange"?: (event: CustomEvent<string>) => void;
         /**
           * Emitted when a suggestion is selected. The `detail` property of the `CustomEvent` will contain the selected suggestion.
          */
-        "onSelected"?: (event: CustomEvent<any>) => void;
+        "onDsoSelect"?: (event: CustomEvent<Suggestion>) => void;
         /**
           * Whether the previous suggestions will be presented when the input gets focus again.
          */
         "suggestOnFocus"?: boolean;
+        /**
+          * The suggestions for the value of the slotted input element
+         */
+        "suggestions"?: Suggestion[];
     }
     interface DsoBadge {
         "status"?: 'primary' | 'success' | 'info' | 'warning' | 'danger';