+++ packages/sources/src/components/autosuggest/autosuggest.args.ts
@@ -1,30 +1,23 @@
-import { HandlerFunction } from "@storybook/addon-actions";
+import { HandlerFunction } from '@storybook/addon-actions';
 
-import { ArgTypes } from "../../stories-helpers";
-
-import { fetchSuggestions } from "./autosuggest.content";
-import { Autosuggest } from "./autosuggest.models";
+import { ArgTypes } from '../../stories-helpers';
 
 export interface AutosuggestArgs {
   suggestOnFocus: boolean;
-  onSelected: HandlerFunction;
+  onSelect: HandlerFunction;
+  onChange: HandlerFunction;
 }
 
 export const autosuggestArgTypes: ArgTypes<AutosuggestArgs> = {
   suggestOnFocus: {
     control: {
-      type: "boolean",
+      type: 'boolean',
     },
   },
-  onSelected: {
-    action: 'onSelected'
+  onSelect: {
+    action: 'onSelect'
+  },
+  onChange: {
+    action: 'onChange'
   }
 };
-
-export function autosuggestArgsMapper(a: AutosuggestArgs): Autosuggest {
-  return {
-    fetchSuggestions: fetchSuggestions,
-    onSelected: a.onSelected,
-    suggestOnFocus: a.suggestOnFocus,
-  };
-}